Move externally developed Sphinx extensions to ``astropy_helpers.extern``
