+++ astropy_helpers/extern/automodapi/templates/autosummary_core/class.rst
