+++ astropy_helpers/extern/numpydoc/numpydoc.py
